refactor: rateパッケージの導入
