+++ dev/20190701/k2works/api/rental/src/main/java/rental/Movie.java
@@ -1,5 +1,7 @@
 package rental;
 
+import rental.rate.Rate;
+
 class Movie {
   String title;
   Rate movieType;
