+++ dev/20190701/k2works/api/rental/src/main/java/rental/Rental.java
@@ -1,5 +1,8 @@
 package rental;
 
+import rental.rate.DaysRented;
+import rental.rate.Rate;
+
 class Rental {
   Movie movie;
   DaysRented daysRented;
