+++ dev/20190701/k2works/api/rental/src/main/java/rental/rate/DaysRented.java
@@ -1,9 +1,9 @@
-package rental;
+package rental.rate;
 
 public class DaysRented {
   int value;
 
-  DaysRented(int value) {
+  public DaysRented(int value) {
     this.value = value;
   }
