+++ dev/20190701/k2works/api/rental/src/main/java/rental/rate/NewRelease.java
@@ -1,4 +1,4 @@
-package rental;
+package rental.rate;
 
 public class NewRelease implements Rate {
   int amountPerDay = 300;
