+++ dev/20190701/k2works/api/rental/src/main/java/rental/rate/OldRelease.java
@@ -1,9 +1,9 @@
-package rental;
+package rental.rate;
 
 public class OldRelease implements Rate {
   MovieType movieType;
 
-  OldRelease(MovieType movieType) {
+  public OldRelease(MovieType movieType) {
     this.movieType = movieType;
   }
