+++ dev/20190701/k2works/api/rental/src/main/java/rental/rate/Rate.java
@@ -1,4 +1,4 @@
-package rental;
+package rental.rate;
 
 public interface Rate {
   int amount(DaysRented daysRented);
