+++ dev/20190701/k2works/api/rental/src/main/java/rental/rate/package-info.java
@@ -1,0 +1,1 @@
+package rental.rate;
